fix errors tests

+++ src/Repository/ReferralsProgramViewRepository.php
@@ -24,7 +24,7 @@ class ReferralsProgramViewRepository extends EntityRepository
             ->getResult()
         ;
 
-        return (int) count($results);
+        return count($results);
     }
 
     public function findMonthReferralsByCustomer(CustomerInterface $customer, \DateTimeInterface $dateTime): int
@@ -39,6 +39,6 @@ class ReferralsProgramViewRepository extends EntityRepository
             ->getResult()
         ;
 
-        return (int) count($results);
+        return count($results);
     }
 }
